#19 Adiciona maps e rota de entregadores proximos

+++ chegaja_frontend/lib/models/deliveryman/deliveryman.dart
@@ -4,7 +4,7 @@ class Deliveryman {
   String? telefone;
   String? nome;
   String? cpf;
-  int? pesoMaximo;
+  double? pesoMaximo;
   int? id;
 
   Deliveryman({
